Address PR review
Remove broken transaction checks in IssuerFlowTest, and note what should be there. The
expectation framework seems to have issues in this test, but the same functionality is covered
in BankOfCordaRPCClientTest where it works correctly, so considering this an issue with the test
and sufficient to cover in the integration tests.

Remove old debugging code from BankOfCordaWebApi